requirements.txt
streamlit>=1.28.0
pandas>=1.5.0
PyPDF2>=3.0.0
google-generativeai>=0.3.0
openpyxl>=3.1.0